Split Radarr models out into multiple files.
