+++ crates/webhook_api/src/models/radarr/mod.rs
@@ -2,55 +2,17 @@
 //! These models are based on the source on Github.
 //! Source: https://github.com/Radarr/Radarr/tree/627ab64fd023269c8bedece61e529329600a3419/src/NzbDrone.Core/Notifications/Webhook
 
+mod movie;
+mod movie_file;
+mod release;
+mod remote_movie;
+
 use crate::models::common::ArrHealthCheckResult;
-use chrono::NaiveDate;
+pub use movie::RadarrMovie;
+pub use movie_file::RadarrMovieFile;
+pub use release::RadarrRelease;
+pub use remote_movie::RadarrRemoteMovie;
 use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct RadarrMovie {
-    pub id: u64,
-    pub title: String,
-    pub file_path: Option<String>,
-    // yyyy-MM-dd as per https://github.com/Radarr/Radarr/blob/627ab64fd023269c8bedece61e529329600a3419/src/NzbDrone.Core/Notifications/Webhook/WebhookMovie.cs#L25
-    pub release_date: Option<NaiveDate>,
-    pub folder_path: Option<String>,
-    pub tmdb_id: Option<u32>,
-    pub imdb_id: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct RadarrRemoteMovie {
-    pub title: String,
-    pub year: Option<u32>,
-    pub tmdb_id: Option<u32>,
-    pub imdb_id: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct RadarrRelease {
-    pub quality: Option<String>,
-    pub quality_version: Option<u32>,
-    pub release_group: Option<String>,
-    pub release_title: Option<String>,
-    pub indexer: Option<String>,
-    pub size: Option<u64>,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct RadarrMovieFile {
-    pub id: u64,
-    pub relative_path: String,
-    pub path: String,
-    pub quality: Option<String>,
-    pub quality_version: Option<u32>,
-    pub release_group: Option<String>,
-    pub scene_name: Option<String>,
-    pub size: Option<u64>,
-}
 
 /// Represents the various webhooks that Radarr can send. The type of webhook is determined by
 /// the webhook's `eventType` property in the body.
